Added feature: Draw tmx maps
@QuakersOat A tmx file is just a reference file, telling you what tile
to put where. You still have to include the original image textures
themselves (in pipeline). Also, we don't include the tmx in pipeline, we
leave it as a separate file and add it to visual studio.

I also added some functionality to the TiledSharp example
(https://github.com/Temeez/TiledSharp-MonoGame-Example), by allowing for
different sized tiles as well as multiple layers.