Fixed spritesheet bleeding
If the view camera for the game is moving around, our drawings actually
get translated based on the position of the camera. However, if the
camera is at at a non-integral position, this transform can throw off
some of our pixel perfect drawings. In sprite sheets, this is especially
bad because this means one sprites image can bleed into another's. There
are two ways to fix this:

1. Always use an integral transform for the camera. (The downside to
this is that it limits what we can do for the camera)
2. Pad spritesheet with blank space (which we have already done), and
write code to select rectangle from spritesheet which does not border
other sprites (leave a gap on the sides of the selection).