Fleshed out core component entity design
Added class scaffolds for components and systems related to player
movement, object rendering, and physics simulation. Next step is to
start implementing these classes.

Definitely read this tutorial to understand what components and systems
are: http://www.richardlord.net/blog/why-use-an-entity-framework

In the context of the above, Artemis is the engine.

Additionally, read https://farseerphysics.codeplex.com/documentation to
understand what fixtures and worlds are in farseer, and read
http://thelinuxlich.github.io/artemis_CSharp/ to better understand
Artemis.

Drop Cave a message on IRC if anything in unclear or if there are any
mistakes.